perf(search): let the reductions table say it has no negative entry
reductions[i] is int(2872 / 128.0 * log(i)) -- a scaled logarithm, 0 through 124
across the whole table, and never negative. As int32_t that is a fact the
compiler cannot use, so the `/ 512` in reduction_of has to allow for a negative
dividend and pays the round-toward-zero correction on every move the search
reduces. uint16_t carries the range instead: both factors of the product are
then provably non-negative and the shift stands alone.

The table also halves, 1024 bytes to 512. Nothing here measures that.

The type is the whole change. search_fill_reductions is the only writer, the
table is filled once from a pure function of the index, and the three
declarations that name the element type move with it.

Ported from ../Stockfish branch refish, `perf(engine): let the reductions table
say it has no negative entry`. That branch REBASES, so its SHAs rot; the subject
is the durable citation. It reads -0.0477% on its own instruction budget, one
compiler wide.

MEASURED HERE:

  ./build.sh signature                2516158 nodes            exit 0
  ./build.sh perf-budget   26072227818 -> 26061980620 instr    exit 0
                                        -10247198   -0.0393%

  Which lands within a hundredth of a point of the sibling's number, on a
  different compiler lane and a different bench -- one of the few in this
  campaign that transfers at its advertised size.

./build.sh parity: all gates passed, exit 0.

No functional change.

+++ src/engine/search/search_common.c
@@ -238,16 +238,22 @@ int singular_triple_margin(
 
 // ---- Step 16 / 17 ------------------------------------------------------
 
-void search_fill_reductions(int32_t *reductions, size_t count) {
+// The entries are int(2872 / 128.0 * log(i)) -- a scaled logarithm, 0 through 124
+// across the whole table and never negative. u16 is what says so: as int32_t that is a
+// fact no compiler can use, so the `/ 512` in reduction_of has to allow for a negative
+// dividend and pays the round-toward-zero correction on every move the search reduces.
+// Both factors of the product below are then provably non-negative and the shift stands
+// alone. The table halves too, 1024 bytes to 512; nothing here measures that.
+void search_fill_reductions(uint16_t *reductions, size_t count) {
     for (size_t i = 1; i < count; ++i)
-        reductions[i] = (int32_t) (2872.0 / 128.0 * log((double) i));
+        reductions[i] = (uint16_t) (2872.0 / 128.0 * log((double) i));
 }
 
-const int32_t *search_reductions_table(void) {
+const uint16_t *search_reductions_table(void) {
     // Fill on first use. The values are a pure function of the index, so a benign
     // race would write byte-identical entries; first use is single-threaded anyway
     // (the main thread's per-worker setup runs before siblings start).
-    static int32_t table[MAX_MOVES];
+    static uint16_t table[MAX_MOVES];
     static bool filled = false;
     if (!filled) {
         search_fill_reductions(table, MAX_MOVES);
@@ -256,7 +262,7 @@ const int32_t *search_reductions_table(void) {
     return table;
 }
 
-int reduction_of(const int32_t *reductions,
+int reduction_of(const uint16_t *reductions,
                  int depth,
                  int move_number,
                  int delta,
